+++ src/plugins/piral-forms/src/withForm.test.tsx
@@ -6,20 +6,12 @@ import * as piralCore from 'piral-core';
 import * as useForm from './useForm';
 import * as usePromise from './usePromise';
 import { describe, it, expect, vitest } from 'vitest';
-import { MemoryRouter } from 'react-router-dom';
 import { render } from '@testing-library/react';
 import { withForm } from './withForm';
 
 vitest.mock('piral-core');
 vitest.mock('./useForm');
 vitest.mock('./usePromise');
-
-const mountWithRouter = (node, url = '/') =>
-  render(
-    <MemoryRouter initialEntries={[url]} initialIndex={0}>
-      {node}
-    </MemoryRouter>,
-  );
 
 const StubComponent: React.FC<{ data: any }> = () => <div role="component" />;
 StubComponent.displayName = 'StubComponent';
@@ -32,6 +24,9 @@ ErrorComponent.displayName = 'ErrorComponent';
 
 (piralCore as any).RegisteredErrorInfo = ErrorComponent;
 (piralCore as any).RegisteredLoadingIndicator = LoaderComponent;
+(piralCore as any).useGlobalStateContext = vitest.fn(() => ({
+  navigation: {},
+}));
 
 describe('withForm Module', () => {
   it('shows error component if nothing is loading and no data is available', () => {
@@ -47,7 +42,7 @@ describe('withForm Module', () => {
     (useForm as any).useForm = usedForm;
     (usePromise as any).usePromise = usedPromise;
     const Component: any = withForm(StubComponent, options);
-    const node = mountWithRouter(<Component />);
+    const node = render(<Component />);
     expect(node.getAllByRole('error').length).toBe(1);
   });
 
@@ -64,7 +59,7 @@ describe('withForm Module', () => {
     (useForm as any).useForm = usedForm;
     (usePromise as any).usePromise = usedPromise;
     const Component: any = withForm(StubComponent, options);
-    const node = mountWithRouter(<Component />);
+    const node = render(<Component />);
     expect(node.getAllByRole('component').length).toBe(1);
   });
 
@@ -81,7 +76,7 @@ describe('withForm Module', () => {
     (useForm as any).useForm = usedForm;
     (usePromise as any).usePromise = usedPromise;
     const Component: any = withForm(StubComponent, options);
-    const node = mountWithRouter(<Component />);
+    const node = render(<Component />);
     expect(node.getAllByRole('loader').length).toBe(1);
   });
 
@@ -102,7 +97,7 @@ describe('withForm Module', () => {
     (usePromise as any).usePromise = usedPromise;
     const options: any = { emptyData: {}, loadData };
     const Component: any = withForm(StubComponent, options);
-    mountWithRouter(<Component />);
+    render(<Component />);
     expect(usedPromise).toHaveBeenCalledTimes(1);
     expect(loadData).toHaveBeenCalledTimes(1);
   });
@@ -124,7 +119,7 @@ describe('withForm Module', () => {
     (usePromise as any).usePromise = usedPromise;
     const options: any = { emptyData: {}, loadData };
     const Component: any = withForm(StubComponent, options);
-    mountWithRouter(<Component />);
+    render(<Component />);
     expect(usedPromise).toHaveBeenCalledTimes(1);
   });
 });