+++ src/plugins/piral-forms/src/withForm.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
-import { withRouter, RouteComponentProps } from 'react-router-dom';
-import { RegisteredLoadingIndicator, RegisteredErrorInfo } from 'piral-core';
+import { RegisteredLoadingIndicator, RegisteredErrorInfo, useGlobalStateContext } from 'piral-core';
 import { useForm } from './useForm';
 import { usePromise } from './usePromise';
 import { InputFormOptions, FormProps } from './types';
@@ -9,15 +8,16 @@ export function withForm<TFormData, TRequiredProps, TProps extends TRequiredProp
   Component: React.ComponentType<TProps & FormProps<TFormData>>,
   options: InputFormOptions<TFormData, TRequiredProps>,
 ): React.FC<TProps> {
-  const FormView: React.FC<TProps & RouteComponentProps & { initialData: TFormData }> = (props) => {
-    const formProps = useForm(props.initialData, props.history, options);
+  const FormView: React.FC<TProps & { initialData: TFormData }> = (props) => {
+    const { navigation } = useGlobalStateContext();
+    const formProps = useForm(props.initialData, navigation, options);
     return (
       <form onSubmit={formProps.submit}>
         <Component {...props} {...formProps} />
       </form>
     );
   };
-  const FormLoader: React.FC<TProps & RouteComponentProps> = (props) => {
+  const FormLoader: React.FC<TProps> = (props) => {
     const { loadData, emptyData } = options;
     const { loading, data, error } = usePromise(() =>
       typeof loadData !== 'function' ? Promise.resolve(emptyData) : loadData(props),
@@ -31,6 +31,7 @@ export function withForm<TFormData, TRequiredProps, TProps extends TRequiredProp
       return <RegisteredErrorInfo type="form" error={error} />;
     }
   };
+  FormLoader.displayName = `withForm(${Component.displayName ?? 'Component'})`;
 
-  return withRouter(FormLoader) as any;
+  return FormLoader;
 }